+++ src/Asset/Adaptor/AbstractAdaptor.php
@@ -79,4 +79,10 @@ abstract class AbstractAdaptor
     {
         return self::$_extensions[self::$type];
     }
+
+    /**
+     * @param $sources
+     * @return mixed
+     */
+    abstract public static function compile($sources);
 }