+++ src/Asset/Compiler.php
@@ -36,15 +36,31 @@ class Compiler
 
     /**
      * @param $files
+     * @param null $postfix
      * @return string
      */
-    public function compile($files)
+    public function compile($files, $postfix = null)
     {
         Registry::flush();
         $files = is_array($files) ? $files : [$files];
         foreach ($files as $file) {
-            Registry::push($file);
+            Registry::push($this->_config, $file);
         }
-        return Registry::compile($this->_config);
+        return Registry::compile($this->_config, $postfix);
+    }
+
+    /**
+     * @param $file
+     * @return string
+     */
+    public function manifest($file)
+    {
+        $manifest   = new Manifest($this->_config, $file, $this);
+        $files      = $manifest->compile();
+        array_unshift($files, $file);
+        return $this->compile(
+            $files,
+            pathinfo($file)['filename']
+        );
     }
 }