+++ src/Asset/File/Registry.php
@@ -10,6 +10,7 @@ namespace Asset\File;
 
 use \Asset\Config;
 use \Asset\Cache;
+use \Asset\Finder;
 
 /**
  * Class Registry
@@ -23,35 +24,54 @@ class Registry
     private static $_files = [];
 
     /**
-     * @param $file
+     * Files hash
+     * @var array
      */
-    public static function push($file)
-    {
-        $ext = self::_getExt($file);
-        $adaptor = '\\Asset\\Adaptor\\' . ucfirst($ext);
-        self::$_files[$adaptor::type()][] = new Virtual($file, $ext);
-    }
-
-    /**
-     * @param $name
-     * @return mixed
-     */
-    private static function _getExt($name)
-    {
-        $name = explode('.', $name);
-        return end($name);
-    }
+    private static $_existsFiles = [];
 
     /**
      * @param Config $config
+     * @param $file
+     */
+    public static function push(Config $config, $file)
+    {
+        if (!realpath($file) && $config->get(Config::BASE_PATH)) {
+            $file = $config->get(Config::BASE_PATH) . DIRECTORY_SEPARATOR . $file;
+        }
+        $finder = new Finder($file);
+        $files  = $finder->search();
+        foreach ($files as $f) {
+            self::_push($f);
+        }
+    }
+
+    /**
+     * @param $file
+     */
+    private static function _push($file)
+    {
+        $ext = Virtual::ext($file);
+        $adaptor = '\\Asset\\Adaptor\\' . ucfirst($ext);
+
+        if (!in_array(md5($file), self::$_existsFiles)) {
+            self::$_existsFiles[] = md5($file);
+            self::$_files[$adaptor::type()][] = new Virtual($file, $ext);
+        }
+    }
+
+
+    /**
+     * @param Config $config
+     * @param null $postfix
      * @return string
      */
-    public static function compile(Config $config)
+    public static function compile(Config $config, $postfix = null)
     {
         $result = '';
         $cache  = new Cache($config);
         foreach (self::$_files as $files) {
             if ($cache->compare($files)) {
+                if ($postfix) { $cache->setPostfix($postfix); }
                 $cache->set($files);
             }
             $result .= $files[0]->serialize($cache->url($files)) . "\n";